Vulkan: figured out the problem; a good solution will be tricky
Command buffers need to pick a framebuffer to draw to. We have command buffers exist per frame-in-flight, but framebuffers are per swap chain image. This means that if a frame's command buffer is built for one swap chain image, but then the frame is asked to draw to another swap chain image, it has to rebuild the command buffer; failing to do so gets us the synchronicity issues (the error message involved the image we're presenting being in the wrong layout, because we were drawing to it and thus resetting it while it was waiting to be presented).

If we don't want to always have to rebuild command buffers every frame (with all the draw calls that that implies), then we need to make command buffers exist per framebuffer and therefore per swap chain image, rather than per frame-in-flight.

Of course, we use a bitmask on rendered_meshes to know what command buffers they need updates for (currently when adding a new mesh, but if we change a mesh's vertices or triangles this would also apply), so we need one bit per command buffer or group of command buffers. If we make command buffers exist per swap chain image, then that means we need an unpredictably-sized bitmask for rendered meshes. No matter what bitmask size we pick, we can't guarantee that some future GPU/OS combination won't mandate a number of swap chain images greater than that number of bits.

Loaded textures need one bit per descriptor set, but descriptor sets are used by command buffers. I suspect, then, that if command buffers are made per swap chain image, then descriptor sets will need to be as well -- which will then require a variably-sized bitmask for loaded textures as well.

I'd prefer for as many resources as possible to be per frame-in-flight in order to reduce GPU resource usage (e.g. from duplicating descriptors/buffers), but the bitmask issue is also a concern.

We may need to find an alternate way to represent the state that the bitmask represents: for "alive" scene assets, which command buffers (meshes) and descriptor sets (textures) need the asset updated; and for scene assets pending deletion, which command buffers and descriptor sets are still using the asset and need to be updated to no longer use it, so that we can safely delete it. We could potentially use a numeric range instead, I think.