Vulkan refactor: it now works!
Most of the functionality that used to be per frame-in-flight is now per swap chain image, to avoid the issues we were having with command buffers and framebuffers. To work around the difficulties with the swap chain image count being unpredictable, we now track meshes' and textures' "dirty" states using numeric ranges instead of bitmasks.

Our descriptor pool now allows the freeing of descriptor sets; this allows us to avoid having to always have to destroy and recreate the pool itself when rebuilding the swap chain. (Otherwise, there'd be no way to free any preexisting descriptor sets except by destroying the pool, so they'd leak.)

We now properly initialize and update the global scene state (i.e. the camera and FOV), and supply that information when drawing. There's a bug where the window has the wrong FOV when initially created, however.

DKVulkanView now disables its paint engine, which is needed to avoid flickering when rendering and resizing.

The render window can now be configured to use DKVulkanView instead of the first renderer, and is currently so configured. It works well, though as noted above we have an FOV bug initially. Rendering (but not program performance) also hitches when (re)building a swap chain while multiple renderers are on-screen.