Vulkan: raycasts at the mouse position now work
The `rendered_mesh` object now has a local bounding sphere usable for quickly doing early-out on raycasts.

Added `surface:renderer::object_index_at`, which does a raycast from the camera at the specified viewport coordinates and returns the index of the nearest hit object.

The render window now has a status bar, and will display an object index there if you click on an object. DKVulkanView emits a signal to facilitate this.