Vulkan: preliminary keyboard camera controls
Moved raycast logic from the `surface_renderer` to a member function on `rendered_mesh`, as that will be more useful.

Implemented keyboard controls for the camera -- currently all hardcoded. We poll for input state. We can discard inputs that are detected after polling was interrupted for a suspiciously long time (current threshold is three seconds). We also track when each key went down, so theoretically we could implement acceleration, distinctions between press and hold actions, and other systems.