Vulkan: camera controls: keybinds now stored centrally, no longer VK codes
We now store 3D renderer keybinds in a singleton, DovahKit3DControls, so that multiple different renderers can refer to it. Keybinds are defined as cobb::qt::key instances.

Forgotten changes from a previous commit: the cobb::qt::key type, intended to represent a keybind in a platform-independent way, but with platform-specific mappings available as and when needed.

THe cobb::qt::keybind struct can now be initialized from a keycode or glyph, and have the appropriate virtual keycode and scancode found at run-time.

Forgotten changes from a previous commit: cobb::vector3 can now be used with "vector-like" objects more easily.