XInput improvements; camera control work
XInput: implemented aim smoothing. For each stick, we track a "real" position and a "simulated" position. The simulated position will "chase" the real position with a maximum speed (unless the "real" position is moving back toward the stick center). This adds an inertia when moving the joystick outward, yet keeps the joystick responsive when moving it back to neutral.

DKVulkanCameraUpdate now stores the elapsed time as a double, and now stores roll/pitch/yaw for the camera rotation.

The vulkanDK::scene object now stores the camera's rotation and position independently and reconstructs the matrix for it as necessary, via vulkanDK::scene::update_camera.

Attempted to rewrite the logic for vulkanDK::scene::adjust_camera. Unfortunately, although the camera can now be moved properly, attempting to rotate the camera instead rotates the scene about (0, 0, 0). Moreover, we remap "object axes" to "camera axes" in adjust_camera when we should probably be doing that in update_camera. In other words, camera behavior is now busted.