More attempts at prototyping render window input
I've settled on having a singleton, DK3DInputHandler, track all of the user's key bindings and the states of each key, rather than having each camera controller track key states individually. Each DKVulkanView will coordinate with DK3DInputHandler, but the input handler will only actually dispatch input to whichever view has focus.

The BoundInput struct represents a single bound key or function. A "boolean" input is, simply, a button: it is down or up. A "scalar" input provides a numeric range, e.g. a joystick trigger. A "vector" input provides a pair of numeric ranges, e.g. mouse movement; notably, any "vector" control can also be treated as two "scalar" controls.

BoundInputState is probably going to be removed as I continue to redesign DK3DInputHandler. Originally I was checking input states per binding, but this means that we'd be making multiple GetAsyncKeyState calls for a keyboard key involved in multiple binds (e.g. a modifier key like Shift). I'm moving to a design where we poll all VKs just once and track their state (including whether a recently-released key was "tapped" or "held"), and then we can check per-binding whether that state satisfies the binding's requirements.

The design that we eventually want is: for each input device (keyboard/mouse and gamepad), a tree of bindings where nodes variously represent global binds, editor modes (object/landscape/navmesh), modifier keys, and individual inputs. Each binding should specify information about how the input is triggered (e.g. "do this thing while the key is down" or "do this thing when the key is released, if it was held only briefly" or "do this thing when the key is released, if it was held for a little while") as well as an editor function to invoke (e.g. "move camera") and arguments to pass to that function. For any given input, we'd use the matching input deepest into the tree, such that descendant nodes can shadow ancestor nodes.