DKKeyPickerWidget; tuned and clarified cobb::qt::key's purpose
Created DKKeyPickerWidget, inspired by a GIF I saw of Cinema4D's keyboard remapping UI. The widget can be used for keyboard shortcuts or single keys.

The cobb::qt::key struct is now explicitly intended to identify a single key on the keyboard for the sake of keyboard shortcuts and mappings, rather than text input. Qt's APIs are not adequate for this task, so it uses native values. There's an internal documentation file in the DK3D folder which goes over specific reasons and edge-cases.

Added a skeletal DKBoundInputWidget, which I intend to use for more complete input mappings (e.g. specifying what kind of button interaction to react to; support for scalar and vector controls; etc.).