DK3D: testing input remapping
DK3DInputHandler's test UI now works on the function-based binds rather than the hardcoded ones. However, there is currently no way to adjust the parameters for each function, which is particularly an issue when changing a binding type (e.g. button -> scalar).

Added name fields to the editor functions. This was purely a quick hack for being able to show function names in the test UI. It's not a good way to associate names with these functions, and it's annoying to maintain; hopefully I'll be able to find a better way when I get to actually doing a proper UI.

Fixed `turn_camera`; it was accidentally disabling scale-by-delta for movement when turning was assigned to a tap-key or hold-key.