Work on 3D input remapping
Numerous fixes to the internals for 3D input remapping, including the `option_union` for tool options.

The UI for remapping 3D keybinds now opens properly and is about half-functional (accounting for the fact that we're still handling keybinds as a flat list for now; in the future, we should use a node tree, with the UI relying on a model).