Vulkan: FPS counter: overlay is now visible
Our FPS counter overlay is now visible, though it doesn't actually measure FPS (it always shows zero). Additionally, it doesn't alpha blend onto the scene properly, and it's slightly blurry (probably due to using the same texture sampler as the scene).

Fixes include...

When setting up our Vulkan instance, we now properly indicate that we're using Vulkan API version 1.1. Previously, we indicated version 1.0 but used 1.1 functions; this caused DovahKit to crash when Vulkan was shimmed by RenderDoc.

The FPS overlay's vertex shader now properly converts pixel coordinates to clip space.

The FPS overlay now properly updates its vertex and index buffer: previously, we acted on the assumption that we only needed to update data which actually changed, but in reality we're writing to an empty staging buffer, so we have to update everything... or almost everything: if we allocate the staging buffer to only include vertices (not indices), then we only need to update the vertices. This means that indices only need to be set once, when the geometry is first created -- and speaking of which, they had the wrong vertex order; they were clockwise and are now counterclockwise, so they're now frontfaces instead of backfaces.

Fixed cobb::enum_flags::empty. Added cobb::enum_flags::test_any_of<...>().

The Debug configuration for DovahKit now runs WinDeployQT, so that the debug build can run standalone outside of VS (necessary for RenderDoc hooking).