FPS counter is now done
The FPS counter is now functional, and uses a cumulative moving average to display frames per second. I implemented this in an attempt to see more "stable" frame rate values, but they still fluctuate wildly from 10000 FPS to 90000 FPS. I suppose that makes sense given how dead-simple this renderer is.

Swap chain images can now update just the command buffer for the FPS counter, avoiding the need to re-record the scene command buffer whenever the counter changes (i.e. every frame).

The surface renderer now offers a texture sampler for pixel graphics, with unnormalized coordinates and nearest-neighbor sampling. This is useful for rendering HUD/UI widgets, with UV coordinates in texture space (i.e. [0, w] instead of [0, 1]) and without artifacts stemming from 3D smoothing. The FPS counter uses this sampler (and the fragment shader for it has been updated to match).

Fixed alpha blending for the FPS counter; we were not properly blending the alpha channel itself.

Optimized (I think) the code for omitting leading zeroes in the FPS counter.

The surface renderer now actually measures how long it takes to render a frame. This measurement is performed on the CPU, so you might expect it to only track how long it takes to update descriptors and shader parameters and record command buffers (as opposed to how long it takes for the GPU to actually perform the work). Remember, however, that we're rendering to a swap chain -- that is, we're cycling through a fixed number of images, reusing each, with a fence to ensure that we wait for an image to be fully painted and presented by the GPU before we try working with it on the CPU. If it takes the GPU a noticeably long amount of time to paint a given swap chain image, then when the CPU hits that image again, it'll end up waiting on the GPU -- and our FPS counter will therefore properly indicate that something has slowed down.