Vulkan: FPS counter fixes and improvements
Fixed the logic for displaying the FPS. Previously, it was using the lowest 5 digits without properly checking whether the number was too large, so for example 57547247 FPS would display as 47247 instead of the intended 99999.

The FPS counter no longer uses a temporary buffer for updating its mesh. We have to update the mesh data every frame, so it makes the most sense to just keep the staging buffer around indefinitely, instead of constantly creating and destroying new ones. Of course, you can't perform buffer copies during a render pass, so `swap_chain_image` has to trigger the copy and *then* start a render pass and trigger the draw call.