Dovahscript: fixed raster resource loading; raster:levels(...)
The code for creating a raster resource now double-checks the format of the inbound QImage and converts it if necessary. This handles cases of loading an image in an unusual format (e.g. 8-bit greyscale JPEG).

Added a `raster:levels(...)` API, which mimics the math from the GNU Image Manipulation Program's "Levels" color adjustment. Code exists for a version that uses vectorized intrinsics to the extent possible (which... isn't very much), but it's slower, so we've disabled it.