Optimized raster:levels; VMA; started work on NIFs
The Dovahscript API `raster:levels({})` now just creates a lookup table for large images, which is orders of magnitude faster than all other approaches including vectorized intrinsics. (Though it does make me wonder... what if we use intrinsics to do the lookups?) We went from 400ms to 30ms for massive images.

Abandoned plans to write my own Vulkan memory allocator; used the VMA library instead. This entailed moving a lot of the code for creating buffers out of `surface_renderer` and into `buffer`. This also means that the `memory_heap.h` and `memory_heap.cpp` files are essentially development leftovers for now.

Started work on code to load NIF files. We'll be focusing on the data needed to render NIFs in our Vulkan engine; other data is non-essential for now. Currently, the only available references for the NetImmerse Format are the various versions of nif.xml scattered around the web, along with whatever can be seen in a hex editor and in NifSkope.