Work on NIF loading
The NIF reader now throws an exception if a `read` call fails, and is capable of retaining error information besides; we now have a set of notice codes for NIFs. The `file::read` function uses a try-catch block. The `file_reader` is no longer a subclass of `cobb::generic_reader`.

Wrote a function to parse a `file_version` from a file format name string.

Wrote a `Float16` struct.

Fleshed out `NiGeometryData`'s loader.