Imported NIFs now render their diffuse textures
Very tentative beginning to the work of breaking `DovahKitCore` into multiple singletons, with `dovahkit::subsystems::assets` just forwarding to it for loading game assets.

Created `cobb::singleton_ex` as a variant on `cobb::singleton` which also creates the getters for you.

The surface renderer now loads DDS textures for NIFs when importing them.