BSA load optimizations; fixed SSE BSA load; added SSE NIF support
Significantly sped up the loading of massive BSAs, including the built-ins for Skyrim Special. Instead of reserving vectors and pushing back, we now preemptively resize and then write to the contents.

Fixed the form load process so that it can pull the proper names for Skyrim Special BSAs.

Added support to the renderer and NIF loading code for Skyrim Special NIFs.

BSVertexDataSSE now stores its vertex colors as a NiColorA, consistent with NiGeometryData, for easier handling by our renderer. Additionally, the loading of bitangent Y/Z components has been corrected (we now actually interpret the byte, and we should get reasonably correct results). Fixed BSVertexDataSSE to actually load its UVs.

Added a codepath to allow for fast loading of a vector of BSVertexDataSSEs.