STAT support; Vulkan shader improvements/fixes; whole-cell import test
Fixes to how we read `NiSkinData`, and optimizations for `NiSkinPartition`.

Modified `surface_renderer::add_nif` so that it can accept a position and a rotation (the latter in lefthanded radians).

Doubled the maximum mesh count for scenes, from 1000 to 2000.

The Render Window now has a test button for importing the contents of a cell into the Vulkan scene on a fire-and-forget basis.

Added more NIF block types. These aren't used in any of the models we're testing; rather, I wanted to have the list of shader properties that NIFs can animate on hand, so I could see whether any properties would be better off if we moved them to push constants. If we can put a piece of per-object data in the push constant rather than in `rendered_mesh::shader_parameters`, we can reduce the size of the all-object-data buffer that each swap chain image needs to rely on.

The fragment shader now always uses tangent space inputs. Previously I had misunderstood exactly why we were using them: they are how we make the normal vector relative to the model's orientation, basically.

The fragment shader now offers control over whether we use Phong or Blinn-Phong shading, assuming I'm properly understanding what I'm learning from.

The vertex shader is now capable of orthogonalizing any TBN vectors it receives. However, this functionality is currently turned off.

Added backend support for loading Static (STAT) forms.