FLOR forms; NiSwitchNodes; fixed LIGH REFR radii in Vulkan
Backend support for FLOR (Flora) forms, and NIF loader and surface renderer support for NiSwitchNodes.

We now handle Bethesda-to-GLM handedness swaps by switching the X- and Y-axis rotations, instead of inverting the first of them that isn't zero. Is that better? I honestly don't know. I suspect we may have to swap X and Y everywhere else to match.

Fixed the handling of LIGH REFRs' radii when importing them into the Vulkan scene.