Directional sun shadows almost work
The renderer is now capable of rendering shadows, but it seems like the shadows clip out in weird ways as the camera moves around. We should use RenderDoc to investigate.

Changes made to get us to this point include...

Deleted `scene_shadow_state`, since the same matrix has to be present in `scene_global_state` anyway. The latter is now passed to the shadow depth shader.

Fixed creation of the shadow depth shader's view/projection matrix.

Removed the shadow depth material's fragment shader, as it's unneeded.

When calculating shadows, surfaces that test as further than 1.0 on the shadow-relative Z-axis are treated as unshadowed (i.e. surfaces outside the range of the shadow map).

Enabled Vulkan-level depth bias instead of shader-based depth bias. Fixed the shadow map shader to override our defaults for render area, so that it'll always render to the full shadow map and not to the window surface extents.

The frustum debugger can now show the shadow map render frustum.