Fixed inverted shadow depth
Turns out, inverting the shadow depth also means you need to negate the shadow bias.

Inverting the depth accomplishes basically nothing on its own. I was hoping I could find a way to also push the far-depth back to infinity, but that may not be mathematically possible with an orthographic matrix.