Renderer is fixed; OIT is yet to be tested
I was recording my command buffers in a different order than they would be executed, which caused some of the state-keeping on `image_and_view` to be incorrect; this is what caused the image layout transition errors.

Separately, the (unavoidable) use of a void pointer for the new push constant helper function allowed me to make a mistake and pass void** instead of void*, breaking all scene object rendering.