Vulkan: shader variants; double-sided NIF support
A `vulkanDK::shader` can now be given `variant`s, to allow for the rapid and simple creation of additional VkPipelines for things like double-sided versus backface-culled shaders.

We now honor the "double-sided" shader flag on imported NIFs.

Started work on improving how we use exceptions within the renderer. Currently, exceptions are only thrown for the bad stuff, and we make no attempt to handle them; part of changing that will be defining specific exception types for our renderer.