Vulkan: shader includes; fixed the FPS counter; prep for shadows
Reorganized the shaders to make much more thorough use of #includes, including for commonly used structs such as the scene global state and rendered mesh push constant.

Fixed depth bias handling for decal meshes: we need to set an initial depth bias before we try to render anything.

Rendered meshes can now be flagged as not casting or not receiving shadows.

Fixed how we measure frame rates within the renderer. Previously, we pretty much only measured the time from acquiring a swap chain image to making a present call; this meant that we didn't measure the influence of the application's own code (e.g. UI processing) on the CPU-side render code. We now factor that in, so our frame rate no longer displays as 99999 every frame.

Started work on adding the needed resources and code for point-light shadows.

Wrote internal documentation for the overall render process, and for how to measure FPS.