Vulkan: per-SCI resources are now per-FIF
Moved most resources to be per frame-in-flight instead of per swap chain image. This means that we need fewer such resources (saving VRAM), and don't have to destroy and recreate them when rebuilding the swap chain. (That last difference, in particular, required some workarounds for properly updating resource states when resizing the render window.)

Fixed `vulkanDK::shader_module`'s constructor so that it properly remembers its own device, and so that it throws an exception on failure.

Fixed a couple shaders in our renderer that weren't properly storing all of their shader modules, resulting in validation errors on teardown (not destroying the shader modules before destroying the VkDevice).