Vulkan: work on point light shadows
Fixed the definition for the point light shadow pass: it now specifies the right render pass.

Fixed the frame-in-flight draw code for the point light shadow pass: we now specify the correct number of clear values, and are careful not to advance past the last subpass.

Added a light type enum to `rendered_light`, in preparation for modifying our shaders to support shadow-omnis.