Vulkan: work continues on point light shadows
I now use arrays for shadow-related state in `scene_global_state`, both in C++ and in GLSL, which has simplified the code a great deal at the cost of no longer being able to explicitly guarantee alignment for each individual array member. (They still *are* aligned, because any sane compiler will align them properly given the object sizes involved, but there's no longer an explicit guarantee.)

Fixed the light space matrices for the "negative" direction for each shadow-caster.

The vertex shader now precomputes the tangent-space light directions, and the light-space vertex positions, for all four shadow-casters. This should be more efficient than having the fragment shader compute the information repeatedly. To facilitate this change, the vertex shader needs access to the light array, and needs to know the array length; ergo it now needs the same specialization constant as the fragment shader core.

In the fragment shader core, we now properly invert `shadow` before using it as as multiplier, when applying point light shadows. That is: we now multiply a lightness value (the opposite of a darkness value) in.

When setting up the light-space projection matrix, we now use `3.141` as our FOV instead of `glm::radians(179.0F)`. Due to arithmetic precision issues, the latter effectively "overflows" and results in calculations equivalent to those done with FOV 180.

All light shadow maps now use the main texture sampler. At some point in the future, I should configure a custom sampler for them, but I don't think that's the cause of my current issues. Removed the individual samplers per shadow map.