Vulkan: more failing to implement point shadows
Fixed resolution on the light shadow map shader.

Added constexpr bools to allow me to disable front-face culling for shadow maps. It's disabled on the maps for placed lights, which is needed for correct shadowing behavior. I'll probably end up having to disable it for the sun as well, once we get around to doing worldspaces.

Blind trial and error random tinkering...