Vulkan: work on shadow cubemaps
Fixed struct layout for a few buffers, including the point light buffer: they needed to use `std430` on the GLSL side of things, and needed padding on the C++ side of things.

Removed PCF code from GLSL `calc_point_shadow` for now, since it wasn't actually doing anything. In the future, we could consider offsetting the vector we query the cubemap with in order to achieve a similar effect.

The vertex shader now computes a light distance ratio for the fragment shader: the distance to the light, divided by the light's radius. In theory, this should be directly comparable to values encoded into the shadow cubemap. The unnormalized fragment-to-light vector is still passed for cubemap lookups.

Fixed a few mistakes in how `vulkanDK::scene` generates the shadow map view/projection matrices. Testing in RenderDoc confirms that we're now rendering the shadow maps correctly.