Vulkan: work on shadow mapping
The shadow cubemaps for point lights are now rendered manually, as color attachments with min or max blend functions, instead of relying on a depth buffer. This is allowing me to fit them linearly to the range [0, 1], instead of the weird [0, 0.0001] or [0.999, 1] ranges that the depth buffers have been using, Examination in RenderDoc shows that this is working.

We can now render shadows, but not strictly correctly. Current test cell is [CELL:03B698]RiftenRatway01 -- one of the corner rooms, with all the FX tiles on the ground. For some reason, our shadows form a set of concentric circles, each half shaded. I suspect that our cubemap faces may be switched around somehow.