Vulkan: shadow seam emulation is now working
Added a test button to the render window which spawns a cube inside of another cube, and a shadow-omni light inside of the innermost cube. The idea is that the inner cube will shadow the outer cube, and any Bethesda-esque shadow seam will appear on the inside of the outer cube.

The fragment shader can now pre-check and skip shadow caster calculations when no casters are within range of the fragment (or no casters exist), though I currently have that turned off.

The fragment shader now emulates the "seam" that appears in Bethesda's shadow-omni lights. The vertex shader supplies data for this as well, but I'm not sure that that data (trigonometric calculations) can be properly interpolated from vertex to vertex; right now, the fragment shader does the calcs entirely on its own.

Shadow seam calculations are dependent on the angle from a fragment to a shadow-casting light. The same angle will be useful in the future for implementing hemi lights. As such, we now store an inverse transformation matrix for rendered lights' shader parameters (i.e. it's calculated CPU-side for performance, since apparently GPUs are slow at matrix inversion).

Added a `surface_renderer::add_light` overload which takes light parameters rather than an ObjectReference.

Wrote documentation comments in `light-shadow-depth.frag`.