Vulkan: support for shadow-casting "hemi" lights
Added support for "hemi" lights to the renderer.

Vertex shader outputs use linear interpolation. This is not suitable for the output of functions like atan; it produces incorrect and highly irregular results. As such, all angle-to-caster math for shadow casters is done on the fragment shader; relevant code and fields have been removed from the vertex shader and vertex-to-fragment struct.

Removed unused code from the code for generating shadow-caster view matrices: the FOV is always 90 now, since we use a cubemap in all cases.