Vulkan: reorganizing the process of drawing a frame; cobb::array_concat
Implemented compute-based culling for the shadow casters' shadow maps: we use the casters' radii.

We now separate out preparation of the draw-indirect parameters into its own step, and we've split more of the frame render process into discrete steps as well. Additionally, submission is now handled by the frames in flight as well.

When we send draw-indirect calls, we now check the mesh index array we built when preparing the calls in the first place. This avoids the need to check the predicates we run (e.g. "does this mesh cast a shadow? if not, skip it for the shadow pass") twice per mesh. That said, we currently do this by mapping GPU memory, which is probably not as fast as it could be. I have something planned for that.

Our descriptor group definitions now support the ability to have a single descriptor set layout be used by multiple descriptor sets within a frame in flight. Additionally, some functions now use `config::frames_in_flight_count` instead of taking it as an argument (the latter was a legacy behavior from when things were per swap chain image, with a count we cannot predict).

Removed the debug shaders for shadow maps, especially since they haven't been functional for a while now.

Created a helper function, `cobb::array_concat`. We now use it along with immediately-invoked lambdas to better organize (and reduce repetition in) the descriptor writes done by `surface_renderer`.