Vulkan: wrapping up the frame rendering refactor
Frames in flight are now stored in an array instead of a vector. This means that their teardown code needs to explicitly destroy member buffers, since we can't simply delete them from existence ahead of the renderer anymore.

Removed unused functions (the old frame rendering flow) from `frame_in_flight`.

We now use device-local buffers for indirect draw params and mesh indices, on the assumption that this would result in faster rendering. So far, it... actually hasn't? Or maybe I'm very slightly CPU-bound, and the ways I've optimized things for the GPU cause slowdowns for the CPU in the Debug build. Not sure.