Vulkan: shader module reuse; teardown fixes
Added functionality to allow the reuse of shader modules, instead of loading multiple copies of one. As a bonus, this also makes the load calls in `surface_renderer` shorter.

Shader modules are now able to skim through the loaded SPIR-V bytecode and check what shader stages are supported by what entry points. We can use this for sanity tests in Debug.

Fixed teardown for shaders, frames in flight, and descriptor set layouts.