Vulkan: updates to internal documentation
Wrote internal documentation for how our compute shader culling works, and for potential future improvements.

Figured out the source for the math we use to compute our frustum plane normals, and read the paper and some other stuff on the web to figure out more-or-less how the math actually works. Documented it all in a code comment.

Updated the internal documentation regarding my overview of the basic render process, to account for adding more semaphores and fences to accommodate the use of additional queues (i.e. compute).