Vulkan: fixed the FPS counter
Our FPS counter was broken because each `frame_in_flight` was measuring and averaging the FPS separately. We currently use two frames in flight, so that meant that we essentially had two distinct counters that were each only seeing half of the frames we rendered. I've fixed this by building a class just for FPS counting (no machinery for display) and attaching it to `surface_renderer`; the FPS overlay class now just pulls the value from there, and is concerned with display only.

The FPS counter now actually counts the number of frames that display during each second. This has proven to be the most accurate FPS tracking method, which is a shame given that it's also the least responsive.