Vulkan: started work on fog
When importing a cell via the test button in the Render Window, we now set scene global state parameters relating to the ambient light and fog.

Added fog-related fields to `scene_global_state`, and wrote shader code to use the fog. Doesn't seem to be working, though...

Current test-case is [CELL:161EB]SkyHavenTemple. It seems like we're not rendering any fog at all.