Vulkan: scene_global_state is now std430; fog still broken
We now use a Vulkan extension to allow uniform buffers to use `std430` layout. This is mandatory if you want to have sensible array packing in any such struct... something which `scene_global_state` relies upon. This fixes shader access to fog parameters and the active shadow caster indices.

Continuing to try and get fog working. It's not going well. We're stuck on maximum fog, no visibility, and I don't know why. RenderDoc straight-up crashes when trying to step through fragment shaders, and that's been a problem for a while now, so I can't actually see how the math is going at run-time.