Worldedit: clicking on forms; handler for form deletion
Finished implementing code for `rendered_mesh::owning_nif`; the pointer is properly written, cleared, acted on by destructors, et cetera.

Replaced `surface_renderer::object_index_at` with `surface_renderer::rendered_mesh_at`, which returns a `rendered_mesh_handle`.

DKVulkanView's `renderedMeshClicked` signal now passes a `vulkanDK::rendered_mesh_handle` instead of a mesh index.

Worldedit now reacts to you clicking on any mesh associated with a form, displaying the form's ID and editor ID (if any) in the Render Window status bar. The debug code in the Render Window, to show mesh indices in the status bar when a mesh is clicked, has been removed.

Worldedit now reacts to the deletion of a loaded CELL or REFR. Placeholders have been added for reacting to the modification of a CELL or REFR.