Worldedit: COC fixes
Worldedit's `_load_cell` function should now listen to its `move_camera_to` argument and let that decide whether to set the camera position. We need this in preparation for worldspace support.

When Worldedit jumps to an interior cell, it now places the camera above the COC marker, in a manner roughly consistent with the Creation Kit.