Bounding boxes in the renderer
Created `cobb::array_flatten`, which recursively flattens `std::array`s given a consistent innermost value type. Put `cobb::array_concat` and my concepts and type traits headers under CC0.

Fixed up Worldedit's selection handling. Made a temporary(!) edit to our mesh click handler so that clicking a mesh toggles its selection state.

The Vulkan renderer now has support for `rendered_bounds`: you can tell it to render a bounding box (and the box's pivot) anywhere in the world. This is separate from `rendered_mesh` because selections in Worldedit are per-NIF, not per-mesh. Shaders are told to draw six vertices but no vertex attributes are defined; shaders deduce the correct color and positions from the vertex index. This avoids the need to create and bind vertex and index buffers. Instancing is also used, and this avoids the need for push constants to indicate which bounding box transform to use.

Frames-in-flight use a separate command buffer for bounds, so we can re-record just that buffer when bounds change, without having to re-record all the other draw commands.

Updated the various scene item handles (e.g. `rendered_mesh_handle` to share a common base class, to cut down on code duplication. CRTP is used to avoid cross-type assignment (i.e. the base class is templated per subclass, so subclasses aren't convertible).

With the above changes, Worldedit can add bounding box displays to the scene. Currently, we use a debug size of (128, 128, 128) for all refs. There is a drawback to my current approach: because a single matrix is used for both the bounding box dimensions and their pivot, all bounding boxes require a centered pivot, which isn't true to how things work in the CK. We'll need to adjust things.